WaveTrackAffordanceHandle refactoring
Moved to be reused
Also, disables "clip rename" action when clip details are not visible

===============================================================================

Cherry-picked from Audacity commit 4bed88bf2.
