+++ src/tracks/playabletrack/wavetrack/ui/WaveTrackAffordanceControls.cpp
@@ -30,9 +30,10 @@
 #include "../../../../Project.h"
 #include "../../../../ProjectWindow.h"
 #include "../../../../commands/AudacityCommand.h"
-#include "../../../ui/AffordanceHandle.h"
+
 #include "../../../ui/TextEditHelper.h"
 #include "WaveTrackView.h"//need only ClipParameters
+#include "WaveTrackAffordanceHandle.h"
 
 #include "../../../../ProjectHistory.h"
 #include "../../../../SelectionState.h"
@@ -47,38 +48,6 @@
 
 #include "WaveClipTrimHandle.h"
 
-class WaveTrackAffordanceHandle final : public AffordanceHandle
-{
-    std::shared_ptr<WaveClip> mTarget;
-public:
-    WaveTrackAffordanceHandle(const std::shared_ptr<Track>& track, const std::shared_ptr<WaveClip>& target) 
-        : AffordanceHandle(track), mTarget(target)
-    { }
-
-    Result Click(const TrackPanelMouseEvent& event, SaucedacityProject* project) override
-    {
-        auto affordanceControl = std::dynamic_pointer_cast<WaveTrackAffordanceControls>(event.pCell);
-        Result result = RefreshCode::RefreshNone;
-        if (affordanceControl)
-        {
-            result |= affordanceControl->OnAffordanceClick(event, project);
-            if (!event.event.GetSkipped())
-                return result;
-            event.event.Skip(false);
-        }
-        return result | AffordanceHandle::Click(event, project);
-    }
-
-    UIHandle::Result SelectAt(const TrackPanelMouseEvent& event, SaucedacityProject* project) override
-    {
-        auto& viewInfo = ViewInfo::Get(*project);
-        viewInfo.selectedRegion.setTimes(mTarget->GetPlayStartTime(), mTarget->GetPlayEndTime());
-        
-        ProjectHistory::Get(*project).ModifyState(false);
-        
-        return RefreshCode::RefreshAll | RefreshCode::Cancelled;
-    }
-};
 
 class SetWaveClipNameCommand : public AudacityCommand
 {
